refactor: drop useless action checkout (#1)
Previously, the action checked out the source code. It happens that the action source code is already checked out. We can leverage the github.action_path context to use it as a working directory.

This allows dropping the explicit reference to the action repository name, which used to be problematic.